Networking: Upgrade to netty 4.0.27
Netty has changed considerably and thus the jump from netty 3 to 4
is a major one. Major topics are explained in more detail below,
most of the work was adhering to netty 4 API.

Netty memory allocation

The idea here is not to use the pooled memory allocation but to restore the
behaviour that existed with netty 3, that everything is unpooled and thus jemalloc
is being disabled.

This does not mean, we will not ever switch to a different allocation model, but keep
the current one for simplicity of changes.

Gathering writes

FIXME need to check, netty does this

HTTP pipelining

Ported pipelining to Netty 4 and changed a fair bit. Pipelining does not
support HTTP chunks anymore, which is not a problem in Elasticsearch as
chunks are aggregated to a FullHttpRequest all the time anyway.

Leftover TODOs

* OutOfMemoryExceptions like this happen when running the tests
[2015-04-09 08:40:10,688][WARN ][netty.channel.DefaultChannelPipeline] An exception was thrown by a user handler's exceptionCaught() method while handling the following exception:
java.lang.OutOfMemoryError: Direct buffer memory
        at java.nio.Bits.reserveMemory(Bits.java:658)
        at java.nio.DirectByteBuffer.<init>(DirectByteBuffer.java:123)
        at java.nio.ByteBuffer.allocateDirect(ByteBuffer.java:311)
        at io.netty.buffer.UnpooledUnsafeDirectByteBuf.allocateDirect(UnpooledUnsafeDirectByteBuf.java:108)
        at io.netty.buffer.UnpooledUnsafeDirectByteBuf.<init>(UnpooledUnsafeDirectByteBuf.java:69)
        at io.netty.buffer.UnpooledByteBufAllocator.newDirectBuffer(UnpooledByteBufAllocator.java:50)
        at io.netty.buffer.AbstractByteBufAllocator.directBuffer(AbstractByteBufAllocator.java:155)
        at io.netty.buffer.AbstractByteBufAllocator.directBuffer(AbstractByteBufAllocator.java:146)
        at io.netty.buffer.AbstractByteBufAllocator.ioBuffer(AbstractByteBufAllocator.java:107)
        at io.netty.channel.FixedRecvByteBufAllocator$HandleImpl.allocate(FixedRecvByteBufAllocator.java:38)
        at io.netty.channel.nio.AbstractNioByteChannel$NioByteUnsafe.read(AbstractNioByteChannel.java:117)
        at io.netty.channel.nio.NioEventLoop.processSelectedKey(NioEventLoop.java:511)
        at io.netty.channel.nio.NioEventLoop.processSelectedKeysOptimized(NioEventLoop.java:468)
        at io.netty.channel.nio.NioEventLoop.processSelectedKeys(NioEventLoop.java:382)
        at io.netty.channel.nio.NioEventLoop.run(NioEventLoop.java:354)
        at io.netty.util.concurrent.SingleThreadEventExecutor$2.run(SingleThreadEventExecutor.java:111)
        at java.lang.Thread.run(Thread.java:745)

The above exception is happening in direct memory, so maybe incoming resources are not closed. Needs more investigation.

* Debugging: Run with -Dio.netty.leakDetectionLevel=advanced in tests and also write heaps by specifying -XX:-HeapDumpOnOutOfMemoryError
-XX:HeapDumpPath=/home/Javin/Desktop/oom_dump.hprof

* Check if `EsHttpContentDecompressor` is still needed

* NettyHeader, use an bytebuf that is never released, see Unpooled.unreleasableBuffer()